\name{Biogrid}
\alias{Biogrid}
\alias{PPI.Biogrid}
\docType{data}
\title{Biogrid Network}
\description{
loads an \code{\link{igraph}} object that contains the Biogrid V 2.0.49 PPI .
}
\usage{data(Biogrid)}
\value{
  an \code{\link{igraph}} named \code{PPI.Biogrid}.
}
\source{
http://thebiogrid.org/
}

\examples{
data(Biogrid)
PPI.Biogrid
}
\keyword{datasets}
