\name{geneLocations}
\alias{geneLocations}
\alias{geneLocations.nucleus}
\docType{data}
\title{
Nucleus located genes
}
\description{
This dataset contains a \code{data.frame} containing the set of genes that are located 
in the nucleus "GO:0005634".
}
\usage{data(geneLocations)}

\value{
	\code{data.frame} containing genes located at the nucleus.
}

\examples{
data(geneLocations)
head(geneLocations.nucleus)
}
\keyword{datasets}
