\name{HPRD}
\alias{HPRD}
\alias{PPI.HPRD}
\docType{data}
\title{
HPRD protein interaction Network
}
\description{
loads an \code{\link{igraph}} object that contains the HRPD PPI release 9.
}
\usage{data(HPRD)}

\source{
http://hprd.org/
}

\value{
  an \code{\link{igraph}} object named \code{PPI.HPRD}.
}
\examples{
data(HPRD)
PPI.HPRD
}
\keyword{datasets}
